bsp: images: complement root-FSs with corresponding FIT-images
Currently barebox cares for bundling a kernel with a devicetree and commandline
options via its support for the boot loader specification [1]. In the context of
a verified boot process this will be replaced by utilizing FIT-images [2]. Thus
prepare a corresponding FIT-image recipe for each root-FS recipe.

As the image's largest component is currently the Linux kernel let's embed a
compressed variant to keep things small.

[1] https://uapi-group.org/specifications/specs/boot_loader_specification/
[2] https://fitspec.osfw.foundation/

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>