bsp: switch from bootspec to FIT-images
Add partitions for storing the FIT-images corresponding to system0, system1 &
failsafe, update barebox' bootscripts accordingly and extend RAUC to include
those FIT-images. Finally adjust the recipes that produce root-FSs to respect
the new dependencies.

As an individual WKS file for creation of SD-card images for the imx6-cpu is
introduced the formerly special handling of injecting the barebox binary into
this image separately for the imx8-cpu can now be simplified again.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>