sw: security: add skov-rauc-keys-native
In the course of introducing the signing class let's condition the file-based
BSP-individual keys for signing RAUC-bundles in form of PKCS#11-URIs by using
soft-HSM: this recipe is a provider for "virtual/rauc-signing".

As this is the second provider now, introduce PREFERRED_PROVIDER_virtual/rauc-
signing for all supported platforms chosing the ptx-dev-keys for imx8-cpu and
the conventional file-based keys for the rest.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>