bsp: images,bundles: switch to secure-boot-enabled RAUC-bundles for imx8-cpu
For the imx8-cpu the recipes for root-FSs used to create partitioned WIC-images
ready to be flashed onto the eMMC or an SD-card. Following the verified boot
strategy one needs to safeguard against tampering of the root-FS. This is done
by introducing dm-verity below the filesystem and appending the corresponding
meta data to the ext4-image. To further enhance security an image individual
salt is chosen for dm-verity's meta data.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>