bsp: images: add secure-boot-enabled images for imx8-cpu
Creation of the required ingredients for a secure-boot-chain follows strict
dependencies: at first the root-FS needs to be created and prepared in the form
of an ext4-filesystem-image. Afterwards this image can be strengthened by the
addition of dm-verity metadata. Only then the verity-root-hash is known and can
be injected into the initramfs which is then taken as an ingredient for the
corresponding FIT-image. As a last step the dm-verity protected filesystem-image
and the related FIT-image need to be filled into partitions of a GPT structured
disk-image. This is what is done here: for each root-FS add a new recipe here
that does the remaining integration into a secure-boot-enabled.

NOTE: In OpenEmbedded every image-recipe has an associated root-FS. Although it
is not needed here (at the moment) it needs to be prepared. Hence formally
choose the barebox package to be installed as it has already been built, is
small and entails no further packages as dependencies.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>