bsp: images: switch to pre-populated writable /home partition
All secbootimg-recipes are particularly image-recipes and therefore have an
associated do_rootfs task although the assembled root-FS has not been used for
anything up to now. Exploit that mechanism to pre-populate the writable /home
partition by extracting the original contents of /home from the original
root-FS' tar archive to the secbootimg-recipes' root-FS in a ROOTFS_POSTPROCESS_
COMMAND. This trick allows to simply restore the previous version of /home's
definition in the corresponding WKS-files.

NOTE: The factory-install-image's /home is empty, thus take a shortcut here and
manually create the directory without any contents.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>