bsp, sw: let systemd-repart & systemd-growfs inflate /home
Up to now the /home partition has been expanded to the maximum possible on the
root-device by the factory-install-image during installation. As modern systemd
has systemd-repart to inflate the underlying partition and systemd-growfs to let
the kernel expand the contained filesystem use those instead. Side effects are
that this fixes the location of the alternate GPT header along the way and also
works on SD-card based installations now.

By default systemd-repart would discard all regions on the root-device that are
not part of any partition. This particularly deletes barebox on SD-cards and is
therefore turned off via a drop-in-file for systemd-repart.service (it requires
a backport of commit [1] from systemd v252 which thus can be removed again from
mickledore onwards as then newer systemd releases are included). On eMMCs an
initial discard of the whole storage medium has already been done by the
factory-install-image.

While at it shrink /home's initial size to 64 MiB in the eMMC's WKS-file to get
a smaller WIC-file for flashing.

[1] 5113436b054a ("repart: always honour `--discard=no`")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>