machine: introduce MACHINE_FEATURE habv4
Instead of differentiating via machine overrides let's assign this new
MACHINE_FEATURE to the imx8s-cpu and do things conditionally on that basis.

Currently this is done only in the barebox recipe, but the mechanism can
probably be used elsewhere, e.g. for signing/not signing a FIT image or
utilizing a FIT image at all.