barebox: imx8s-cpu: enable OP-TEE
Adjust these build-time configuration parameters

  enable CONFIG_TEE
  enable CONFIG_OPTEE
  enable CONFIG_PBL_OPTEE
  enable CONFIG_FIRMWARE_IMX8MP_OPTEE

and set the following according to the machine parameters from imx8s-cpu.conf:

  set CONFIG_OPTEE_SIZE=0x02000000  <--  OPTEE_CORE_SIZE + OPTEE_SHMEM_SIZE
  set CONFIG_OPTEE_SHM_SIZE=0x200000  <--   OPTEE_SHMEM_SIZE

Add a check to the recipe to ensure that these manually set values in barebox'
defconfig do not diverge from those defined in the imx8s-cpu's machine
configuration in the future.

Copy the OP-TEE-adjusted version of the TF-A as well as the OP-TEE binary so
that they are available for consumption by barebox' compile task. For the
latter, an additional dependency is added to ensure deployment.