barebox: optimize handling of dependencies
Only the firmwares' deploy tasks need to be completed for barebox' compile task
to run, so explicitly move those dependencies to do_compile (and not to
do_configure, which previously triggered one of the dependencies).

This requires changing the source from STAGING_DIR_TARGET to DEPLOY_DIR_IMAGE
when copying the TF-A's binary blob, since the former is no longer populated
with files from the TF-A anymore. Instead of directly copying the regular file
'bl31-imx8mp.bin', use the symbolic link 'imx8mp-bl31.bin' prepared by the
preceding commit, and dereference it while copying using the '-L' option.