scripts: wks: leave home partition empty and explicitly zero the first kiB
Previously, the home partition directly contained a filesystem (see [1]). Now,
it will be integrity-protected and encrypted before a filesystem is created.
This encryption is done using a SoC-specific trusted key, which is stored as a
sealed trusted key at the beginning of the home partition. Since this key is (a)
SoC-specific and (b) not known offline, the integrity-protection and encryption
can not be prepared beforehand. Thus, create an empty partition now and
explicitly clear the first kiB of data to remove any leftover key material from
previous installations. The filesystem is going to be created on first boot.

[1] 5a2e073b1ffa ("bsp: images: switch to pre-populated writable /home partition")