linux-skov: update to 6.14/customers/skov/imx6/20250815-1 (6.14.11 -> 6.14.11)
Disable caam job ring 2 as it is going to be used by OP-TEE in the secure world
and therefore unavailable to the Linux kernel which runs in the normal world.

Then for all machines, run the normal dance of (exemplarily for the imx8s-cpu)

  $ bitbake -c clean linux-skov
  $ bitbake -c configure linux-skov
  $ bitbake -c devshell linux-skov
    # export CROSS_COMPILE=aarch64-pmc-linux-
    # export KBUILD_OUTPUT=<PATH_TO_BUILDDIR>/tmp/work/imx8s_cpu-pmc-linux/linux-skov/6.14.11/build
    # make menuconfig
    # cp <PATH_TO_BUILDDIR>/tmp/work/imx8s_cpu-pmc-linux/linux-skov/6.14.11/build/.config \
         <PATH_TO_BSPDIR>/meta-skov-cpu-bsp/recipes-kernel/linux/linux-skov/imx8s-cpu/defconfig
    # exit

to clean up the defconfigs, so that one can later focus only on the intended
changes.