openssl: prepare customization of openssl.cnf
In a subsequent commit, OpenSSL's configuration file will be adapted to utilize
OP-TEE's PKCS#11 capabilities. To avoid missing updates to the upstream version,
a check is added that alerts when the customized version needs to be updated
accordingly.

As a first step, this commit only installs the unmodified upstream openssl.cnf,
extracted via

  $ bitbake -c clean openssl
  $ bitbake -c compile openssl
  $ bitbake-getvar -r openssl B
  $ cp <B>/apps/openssl.cnf recipes-connectivity/openssl/openssl/customized_openssl.cnf

from OpenSSL 3.2.4. It is identical to [1].

[1] https://github.com/openssl/openssl/blob/8ac4271233ecec5aa0468a5e7ae38dd642ebf365/apps/openssl.cnf