barebox: imx6-cpu: enable MMCBLKDEV_ROOTARG
Kernels prior to v5.10-rc1 could not guarantee deterministic ordering of MMC
devices. Therefore, the default behavior of barebox 'global.bootm.appendroot'
and blspec 'linux-appendroot' for MMC block devices is to append a
'root=PARTUUID=XYZ' option to the kernel command line. The PARTUUID is derived
from the partition table and is expected to be unique across the system.

This approach can fail if the system is booted with an identical MMC image on
multiple boot media (e.g. SD and USB), where identical PARTUUIDs collide. In
such setups, the only robust alternative used to be an initrd.

Recent kernels honor the numbering derived from 'mmc' aliases, which makes
'root=mmcblkXpN' stable and deterministic as long as barebox and the kernel use
matching alias definitions.