barebox: imx8s-cpu: do not start "platsch" as PID1 on headless systems
Big Dutchman's bd500 variant of the IMX8 board (board 2, rev c, compatible
"imx8mp-skov-revc-bd500") doesn't have the DRM-devices required by "platsch" so
that "platsch" exits with an error condition as a result. Since it is started as
PID1 this leads to a kernel panic. There is no need for a splash screen on a
headless system so just start the default in that case (init=/usr/sbin/init
which is a symbolic link to /usr/lib/systemd/systemd).